added login

+++ src/api/deleteTodo.ts
@@ -1,4 +1,9 @@
 export async function deleteTodo(id: number) {
+
+    if(typeof id !== 'number' || isNaN(id)) {
+        throw new Error('Invalid ID passed to delete')
+    }
+    
     const url = `http://localhost:6213/api/todo/${id}`;
     const response = await fetch(url, {
         method: 'DELETE',
